Rename component folders
