+++ JoyboyCommunity/src/components/_Avatar/styled.ts
