+++ JoyboyCommunity/src/components/_Button/index.tsx
@@ -1,7 +1,7 @@
 import {Pressable, PressableProps, StyleProp, TextStyle} from 'react-native';
 
 import {useStyles} from '../../hooks';
-import {Text} from '../text';
+import {Text} from '../_Text';
 import stylesheet from './styles';
 
 export type ButtonProps = PressableProps & {
