+++ JoyboyCommunity/src/components/index.ts
@@ -1,15 +1,15 @@
-export {Avatar, AvatarImage} from './avatar';
-export {Button} from './button';
-export {Divider} from './divider';
-export {Header} from './header';
-export {IconButton} from './iconbutton';
-export {Input} from './input';
-export {Menu} from './menu';
-export {KeyboardAvoidingView} from './skeleton/KeyboardAvoidingView';
-export {LayoutContainer} from './skeleton/LayoutContainer';
-export {RootScreenContainer} from './skeleton/RootScreenContainer';
-export {RowContainer} from './skeleton/RowContainer';
-export {ScreenContainer} from './skeleton/ScreenContainer';
-export {ScrollableContainer} from './skeleton/ScrollableContainer';
-export {Text} from './text';
-export {TextButton} from './textbutton';
+export {Avatar, AvatarImage} from './_Avatar';
+export {Button} from './_Button';
+export {Divider} from './_Divider';
+export {Header} from './_Header';
+export {IconButton} from './_IconButton';
+export {Input} from './_Input';
+export {Menu} from './_Menu';
+export {KeyboardAvoidingView} from './_Skeleton/KeyboardAvoidingView';
+export {LayoutContainer} from './_Skeleton/LayoutContainer';
+export {RootScreenContainer} from './_Skeleton/RootScreenContainer';
+export {RowContainer} from './_Skeleton/RowContainer';
+export {ScreenContainer} from './_Skeleton/ScreenContainer';
+export {ScrollableContainer} from './_Skeleton/ScrollableContainer';
+export {Text} from './_Text';
+export {TextButton} from './TextButton';
